Fix allocation bug for matrices, compute underlying vector size as |std::size_t|
While I must have decided in the past that row and column sizes could be handled
as |unsigned int| (in practice 32-bits quantities), that still requires the
storage vector to have a size potentially requiring 64-bits, so |std::size_t| is
required for the size computation in the various |Matrix_base| constructors. The
actual address arithmetic for indexing was already done using |std::size_t|.

+++ sources/utilities/matrix.h
@@ -161,9 +161,9 @@ template<typename C> class Matrix_base
   Matrix_base(): d_rows(0),d_columns(0),d_data() {}
 
   Matrix_base(unsigned int m, unsigned int n)
-    : d_rows(m),d_columns(n),d_data(m*n) {}
+    : d_rows(m),d_columns(n),d_data(static_cast<std::size_t>(m)*n) {}
   Matrix_base(unsigned int m, unsigned int n, const C& c)
-    : d_rows(m), d_columns(n), d_data(m*n,c) {}
+    : d_rows(m), d_columns(n), d_data(static_cast<std::size_t>(m)*n,c) {}
 
   Matrix_base (const std::vector<Vector<C> >& b,
 	       unsigned int n_rows); // with explicit #rows in case |b| empty
